Update
Added elements to begin matching the given design. Started to stylize it, resized the hero image, added border-radius to main card and hero-image, will need to adjust the font size and margins, I resized the author-avatar and added a 50% border-radius, added flex to the author-credit section at bottom of card, and added background color to the body.