Update
I was working on getting the JS to play nice with the share button, still trying to figure out how to keep it in the same position when hiding the author/date of publication. I will need to incorporate an if-statement to be able to toggle between hiding and revealing the social media sharing options.